Login Testing - Read Description
Currently testing the Login page for the back end code with static login credentials, it works perfectly fine, but if you enter the wrong details then an error will show telling the user the credentials are wrong.

I have added a "welcome.jsp" file just for testing, although this is not the final version as an updated version will be added soon by the front-end developers.