Fixed error with Mapping for login.jsp and MainController - See Description
After doing a small bit of testing in the morning I managed to run into an error which cause the project to not start at all, after looking into the error I noticed that the error was cause by the mapping parameters which had the values "/login".

This should now be working without an error and allowing users to login with the current static credentials - Email: Admin@f4u.com & Password: Root.